refactor: move PowerMeasure implementations to better package
